+++ aeronet_webtransport/src/client/frontend.rs
@@ -2,6 +2,7 @@ use std::future::Future;
 
 use aeronet::client::ClientState;
 use futures::channel::oneshot;
+use xwt_core::utils::maybe;
 
 use super::{backend, ClientConfig, ClientError, Connected, Connecting, WebTransportClient};
 
@@ -14,7 +15,20 @@ impl WebTransportClient {
         }
     }
 
-    pub fn connect_new(config: ClientConfig) -> (Self, impl Future<Output = ()> + Send) {
+    pub fn disconnect(&mut self) -> Result<(), ClientError> {
+        match self.state {
+            State::Disconnected => Err(ClientError::AlreadyDisconnected),
+            State::Connecting(_) | State::Connected(_) => {
+                *self = Self::disconnected();
+                Ok(())
+            }
+        }
+    }
+
+    pub fn connect_new(
+        config: ClientConfig,
+        target: impl Into<String>,
+    ) -> (Self, impl Future<Output = ()> + maybe::Send) {
         let (send_connected, recv_connected) = oneshot::channel::<Connected>();
         let (send_err, recv_err) = oneshot::channel::<ClientError>();
 
@@ -24,11 +38,27 @@ impl WebTransportClient {
                 recv_err,
             }),
         };
+        let target = target.into();
         let backend = async move {
-            if let Err(err) = backend::start(config, send_connected).await {
+            if let Err(err) = backend::start(config, target, send_connected).await {
                 let _ = send_err.send(err);
             }
         };
         (frontend, backend)
+    }
+
+    pub fn connect(
+        &mut self,
+        config: ClientConfig,
+        target: impl Into<String>,
+    ) -> Result<impl Future<Output = ()> + maybe::Send, ClientError> {
+        match self.state {
+            State::Disconnected => {
+                let (frontend, backend) = Self::connect_new(config, target);
+                *self = frontend;
+                Ok(backend)
+            }
+            State::Connecting(_) | State::Connected(_) => Err(ClientError::AlreadyConnected),
+        }
     }
 }